buildCode: rename format -> build
